Renamed files/variables, changed project structure, added Header to index.html, initial working functionality of search trainees page
