+++ src/app/models/trainee.ts
@@ -4,7 +4,7 @@ export class Trainee {
     firstname: string;
     lastname: string;
     batch_id: string;
+    password: string;
     city_preferences: any[];
     doman_preferences: string[];
-    password: string;
 }
