+++ src/app/searchTraineesComponent/app.searchTraineesComponent.css
@@ -72,10 +72,6 @@ input {
     overflow-y:scroll;
     -webkit-overflow-scrolling: touch;
 }
-
-.navbar.navbar-1 .navbar-toggler-icon {
-    background-image: url('https://mdbootstrap.com/img/svg/hamburger6.svg?color=000');
-}
 .panel {
 	padding: 0px;
 }
