+++ src/app/services/TraineeService.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { Trainee } from './trainee';
+import { Trainee } from '../models/trainee';
 
 @Injectable( {
     providedIn: 'root',
